almost there

+++ Ass2/Scene.h
@@ -11,6 +11,9 @@
 #include "glut.h"
 #include "Vector3f.h"
 #include "Camera.h"
+#include "Spher.h"
+#include "Plane.h"
+
 
 using namespace std;
 typedef Vector3f Point;
@@ -20,17 +23,18 @@ class Scene {
 public:
 
 	Scene();
-	Scene(Point center, Vector3f up, int w, int x, int y, Light* light, Camera* cam);
+	Scene(Point center, Vector3f up, GLfloat w, GLfloat x, GLfloat y, Light* light, Camera* cam);
 	~Scene();
 	void addLight(Light* light);
 	void addObject(SceneObject* object);
+
 
 	Point screenCenter;
 	Vector3f upVector;
 	Vector3f toVector;
 	Vector3f rightVector;
-	int width;
-	int Rx, Ry;
+	GLfloat width;
+	GLfloat Rx, Ry;
 	Light* ambient;
 	Camera* camera;
 	vector<Light*> lights;
